refactoring devicetools module

+++ devicetools.py
@@ -6,6 +6,7 @@ def int_format(sval, frm):
     if not (isinstance(frm, str) and len(frm)):
         frm = 'gddd'
     sign = ''
+    print(f'sval:{sval}, frm:{frm}')
     if isinstance(frm, str) and len(frm):
         if frm[0].lower() == 'g':
             sign = 'g'
@@ -14,67 +15,74 @@ def int_format(sval, frm):
     if sval is None:
         return None
     else:
-        if isinstance(sval, bool):
-            sval = '1' if sval is True else '0'
-        if sval == 'False':
-            sval = '0'
+        try:
+            sval = int(sval)
+        except ValueError:
+            pass
+
+        if isinstance(sval, str):
+            if (sval == '') or (sval.lower()[0] == 'f'):
+                sval = '0'
+            else:
+                sval = '1'
 
         if not isinstance(sval, str):
             sval = str(sval)
 
-        if '.' in sval:
-            sval = str(sval).split('.')
-            sval = sval[0]
-        if sval:
-            if (sign == 'g') and (sval[0] == '-'):
-                sign = '-'
-                sval = sval[1:]
-            else:
-                sign = ''
-            nb = ''
-            for c in sval:
-                if 47 < ord(c) < 58:
-                    nb += c
-            sval = nb
-            if len(sval) > len(frm):
-                sval = sval[len(sval) - len(frm):]
-            sval = sign + sval
-            try:
-                sval = int(sval)
-            except ValueError:
-                sval = None
-            return sval
-        return None
+        if (sign == 'g') and (sval[0] == '-'):
+            sign = '-'
+            sval = sval[1:]
+        else:
+            sign = ''
+        nb = ''
+        for c in sval:
+            if 47 < ord(c) < 58:
+                nb += c
+        sval = nb
+        print(f'sval:{sval}')
+        if len(sval) > len(frm):
+            sval = sval[len(sval) - len(frm):]
+        sval = sign + sval
+
+        return int(sval)
 
 
 def float_format(sval, frm):
-    if not (isinstance(frm, str) and len(frm) and ('.' in frm)):
+    if not (isinstance(frm, str) and len(frm)):
         frm = 'gfff.ff'
 
-    if sval is None:
+    if sval in (None, False, True):
         return None
+    if isinstance(sval, str):
+        while len(sval) and not (47 < ord(sval[-1]) < 58):
+            sval = sval[:-1]
+        if len(sval) == 0:
+            return None
+        try:
+            sval = float(sval)
+        except ValueError:
+            return None
+
     frm = frm.lower().replace('f', 'd')
     ar = frm.split('.')
 
     if not isinstance(sval, float):
-        try:
-            sval = float(sval)
-        except ValueError:
-            pass
+        sval = float(sval)
+
     if isinstance(sval, float):
-        sval = f'{sval:.{len(ar[1])}f}'
+        if len(ar) > 1:
+            sval = f'{sval:.{len(ar[1])}f}'
 
     sar = str(sval).split('.')
+    # print(f'sval:{sval}, sar:{sar}, ar:{ar}')
     d = int_format(sar[0], frm=ar[0])
+    print(f'd:{d}')
     if len(sar) > 1:
         sval = f'{d}.{sar[1]}'
     else:
         sval = f'{d}.0'
-    try:
-        sval = float(sval)
-    except ValueError:
-        return None
-    return sval
+
+    return float(sval)
 
 
 def write_file(filename, dat):
@@ -106,6 +114,8 @@ def read_file(filename):
         Returns:
             data (dict): python dictionary or None when data is't load.
     """
+    if not filename:
+        return None
     try:
         with open(filename, "r") as f:
             dat = f.read()
@@ -126,6 +136,8 @@ def remove_file(filename):
         Returns:
             (bool): result deleting file.
     """
+    if not filename:
+        return False
     try:
         os.remove(filename)
     except OSError:
